+++ pages/page-memorizzazione-sutra.php
@@ -1,0 +1,3 @@
+<div class="cz-sutra-memo-container">
+<?php echo do_shortcode('[cz-sutra-memo]'); ?>
+</div>