+++ parts/footer.php
@@ -1,7 +1,12 @@
 	<footer>
 		<div class="wp-group">
 			<p><?php echo date('Y'); ?>, Cigno Zen</p>
-			<p><a href="<?php echo esc_url(home_url('/contatti')); ?>">Contatti</a> | <a href="<?php echo esc_url(home_url('/privacy-policy')); ?>">Privacy</a></p>
+			<p>
+				<a href="<?php echo esc_url('https://www.instagram.com/cignozen'); ?>">Instagram</a> ・
+				<a href="<?php echo esc_url(home_url('/contatti')); ?>">Contatti</a> ・
+				<a href="<?php echo esc_url(home_url('/privacy-policy')); ?>">Privacy</a> ・
+				<a href="<?php echo esc_url(home_url('/utilita')); ?>">Utilità</a>
+			</p>
 		</div>
 	</footer>
 	<?php wp_footer(); ?>